Flow View buttons to inline
